+++ kinto/core/storage/memory.py
@@ -1,10 +1,7 @@
-import json
 import numbers
 import operator
 import re
 from collections import abc, defaultdict
-
-import ujson
 
 from kinto.core import utils
 from kinto.core.decorators import deprecate_kwargs, synchronized
@@ -16,7 +13,7 @@ from kinto.core.storage import (
     StorageBase,
     exceptions,
 )
-from kinto.core.utils import COMPARISON, find_nested_value
+from kinto.core.utils import COMPARISON, find_nested_value, json
 
 
 def tree():
@@ -27,9 +24,6 @@ class MemoryBasedStorage(StorageBase):
     """Abstract storage class, providing basic operations and
     methods for in-memory implementations of sorting and filtering.
     """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
 
     def initialize_schema(self, dry_run=False):
         # Nothing to do.
@@ -136,11 +130,6 @@ class Storage(MemoryBasedStorage):
     Enable in configuration::
 
         kinto.storage_backend = kinto.core.storage.memory
-
-    Enable strict json validation before saving (instead of the more lenient
-    ujson, see #1238)::
-
-        kinto.storage_strict_json = true
     """
 
     def __init__(self, *args, readonly=False, **kwargs):
@@ -189,7 +178,13 @@ class Storage(MemoryBasedStorage):
         modified_field=DEFAULT_MODIFIED_FIELD,
     ):
         id_generator = id_generator or self.id_generator
-        obj = {**obj}
+
+        # This is very inefficient, but memory storage is not used in production.
+        # The serialization provides the necessary consistency with other
+        # backends implementation, and the deserialization creates a deep
+        # copy of the passed object.
+        obj = json.loads(json.dumps(obj))
+
         if id_field in obj:
             # Raise unicity error if object with same id already exists.
             try:
@@ -202,7 +197,6 @@ class Storage(MemoryBasedStorage):
 
         self.set_object_timestamp(resource_name, parent_id, obj, modified_field=modified_field)
         _id = obj[id_field]
-        obj = ujson.loads(self.json.dumps(obj))
         self._store[parent_id][resource_name][_id] = obj
         self._cemetery[parent_id][resource_name].pop(_id, None)
         return obj
@@ -233,9 +227,13 @@ class Storage(MemoryBasedStorage):
         id_field=DEFAULT_ID_FIELD,
         modified_field=DEFAULT_MODIFIED_FIELD,
     ):
-        obj = {**obj}
+        # This is very inefficient, but memory storage is not used in production.
+        # The serialization provides the necessary consistency with other
+        # backends implementation, and the deserialization creates a deep
+        # copy of the passed object.
+        obj = json.loads(json.dumps(obj))
+
         obj[id_field] = object_id
-        obj = ujson.loads(self.json.dumps(obj))
 
         self.set_object_timestamp(resource_name, parent_id, obj, modified_field=modified_field)
         self._store[parent_id][resource_name][object_id] = obj
@@ -442,7 +440,8 @@ def extract_object_set(
 
 
 def canonical_json(obj):
-    return json.dumps(obj, sort_keys=True, separators=(",", ":"))
+    # We just a predictable serialization so that we just compare strings.
+    return json.dumps(obj, sort_keys=True)
 
 
 def apply_filters(objects, filters):
@@ -586,6 +585,4 @@ def _get_objects_by_parent_id(store, parent_id, resource_name, with_meta=False):
 
 
 def load_from_config(config):
-    settings = {**config.get_settings()}
-    strict = settings.get("storage_strict_json", False)
-    return Storage(strict_json=strict)
+    return Storage()